requirements.txt
streamlit
google-genai
pandas(for storing csv only )